refactor(core): split Security and Agent multi-type files into per-type files
Split ISecurityManager.cs (583 lines) and IAgentOrchestrator.cs (763 lines),
each of which declared multiple public types, into one file per public type
to follow the single-responsibility / one-type-per-file convention:

Security/ : ISecurityManager, ToolApprovalResult, SecurityCheckResult,
            ToolExecutionAuditEntry, UserPermissionSettings, SecurityManager
Agent/    : ISLMService, IAgentOrchestrator, ToolCall, ToolExecutionResult,
            AgentResponse, AgentOrchestrator

No logic changes; namespaces and behavior preserved. Core project builds clean.

Addresses review threads PRRT_kwDOPHDmb85VLj2h and PRRT_kwDOPHDmb85VLj2k.
